refactor Selenium fabric
