+++ src/main/java/executor/service/service/FactoryForDI.java
@@ -8,11 +8,11 @@ import executor.service.service.scenario.ScenarioSourceListener;
 import executor.service.service.scenario.selenium.ScenarioExecutorImpl;
 import executor.service.service.scenario.common.ScenarioSourceListenerImpl;
 import executor.service.service.stepexecutors.selenium.StepExecutionClickCss;
-import executor.service.service.stepexecutors.selenium.StepExecutionClickCssImpl;
+import executor.service.service.stepexecutors.selenium.impl.StepExecutionClickCssImpl;
 import executor.service.service.stepexecutors.selenium.StepExecutionClickXpath;
-import executor.service.service.stepexecutors.selenium.StepExecutionClickXpathImpl;
+import executor.service.service.stepexecutors.selenium.impl.StepExecutionClickXpathImpl;
 import executor.service.service.stepexecutors.selenium.StepExecutionSleep;
-import executor.service.service.stepexecutors.selenium.StepExecutionSleepImpl;
+import executor.service.service.stepexecutors.selenium.impl.StepExecutionSleepImpl;
 import executor.service.service.webdrivers.WebDriverInitializer;
 
 public class FactoryForDI {
