+++ src/main/java/executor/service/service/stepexecutors/selenium/impl/StepExecutionClickCssImpl.java
@@ -1,6 +1,7 @@
-package executor.service.service.stepexecutors.selenium;
+package executor.service.service.stepexecutors.selenium.impl;
 
 import executor.service.model.entity.Step;
+import executor.service.service.stepexecutors.selenium.StepExecutionClickCss;
 import org.openqa.selenium.WebDriver;
 
 public class StepExecutionClickCssImpl implements StepExecutionClickCss {
