+++ src/main/java/executor/service/service/stepexecutors/selenium/impl/StepExecutionSleepImpl.java
@@ -1,9 +1,10 @@
-package executor.service.service.stepexecutors.selenium;
+package executor.service.service.stepexecutors.selenium.impl;
 
 import executor.service.model.entity.Step;
+import executor.service.service.stepexecutors.selenium.StepExecutionSleep;
 import org.openqa.selenium.WebDriver;
 
-public class StepExecutionClickXpathImpl implements StepExecutionClickXpath {
+public class StepExecutionSleepImpl implements StepExecutionSleep {
     @Override
     public String getStepAction() {
         return null;
